Set keyboard frame to below the window when resigning first responder
We want the accessory view at the bottom when the keyboard is undocked.
When undocking apple fires 'correctly' the will hide keyboard
notification, but the frame is where the keyboard will be, when
undocked. We therefor need to intersect it in these cases and return a
rect below the window.