Simplify logic to work with iOS9+
The original implementation was used before Apple moved the keyboard
into a dedicated window and they moved the input accessory view by it
self. Then it was rather easy to get notified when the accessory view
got moved, which meant the keyboard got moved.

Instead of doing this we're now simply using a UIPanGestureRecognizer
and manually moving it. This means the entire input accessory view is
redundant*. Which means its less moving parts and less code. This means
interactive dismissing now works on iOS9+ :tada:.

*: We probably want to reintroduce an input accessory view, to get the
desired behavior when using the keyboard in split-mode.